refactor: add oracle registration to allow different oracles to be used in the system (#116)
* chore: move folders

* build: add oracle id into productId

* chore: add snapshot

* build: merge ChainlinkPricer and oracle into ChainlinkOracle

* build: oracle registration

* chore: lint

* build: remove default oracle in AdvancedMargin

* chore: fix deploy script

* chore: cleanup

* test: test revert when price is un-reported

* test: cannot set price for the future

* test: cover more decimal cases

* chore: lint

* chore: cleanup mockOracle

* chore: remove redundant line in MockEngine

* chore: add engine.isAccountHealthy check to improve coverage

* chore: add test for transfer account to improve coverage

* chore: gas-snapshot

+++ src/core/engines/advanced-margin/AdvancedMarginLib.sol
@@ -15,7 +15,7 @@ import "../../../config/errors.sol";
  */
 library AdvancedMarginLib {
     using TokenIdUtil for uint256;
-    using ProductIdUtil for uint32;
+    using ProductIdUtil for uint40;
 
     /**
      * @dev return true if the account has no short positions nor collateral
@@ -61,7 +61,7 @@ library AdvancedMarginLib {
         uint256 tokenId,
         uint64 amount
     ) internal {
-        (TokenType optionType, uint32 productId, , , ) = tokenId.parseTokenId();
+        (TokenType optionType, uint40 productId, , , ) = tokenId.parseTokenId();
 
         // assign collateralId or check collateral id is the same
         uint8 collateralId = productId.getCollateralId();
